WIP: Split numato tests to multiple files
test_hass_binary_sensor_notification still fails.
